requirements.txt
psycopg2-binary==2.9.11
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.38
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
